Communicate package dependencies from solver to builder precisely.
This should in principle fix quite a few subtle problems that
presented themselves as the solver finding a valid install plan,
but the builder failing.

The reason is that until now, once the solver determined an exact
install plan, only imprecise information would be used to invoke
the builder. In particular, for installed packages, not the exact
instance would be chosen, but only the version would be selected.
In situations where multiple instances of the same package version
are present in the package DB stack, this could lead to the builder
selecting a different instance than intended by the solver, and
as a consequence, most likely to a configure or compile time error.

The current implementation is a bit rough around the edges and
still produces some debugging output, but it seems to work.

There's a new flag

  --package-id

available for the configure phase. Using --package-id, an exact
package identifier can be specified to be used.

I have decided not to use --constraint for this, because
--constraint on the Cabal level only takes dependency constraints,
and there are too many assumptions everywhere in the code that
dependency constraints should just be package names paired with
a version range.