Corner cases of --package-id config flag.
The flag for now should not be inherited by "install". It's not
really intended to be a user-level flag. We could add it to
"install" at a later stage, but then it also has to be treated
correctly for that case, i.e., fed correctly as an additional
constraint into the solver. This is additional work!

Also added filtering rules such that if the external build method
is being used with an older version of Cabal, then "--package-id"
is not going to be passed.